Get a SINIM variable data in a specific year as a data frame

Get a SINIM variable data in a specific year as a data frame

getsinimr(var, year, moncorr = T)

Arguments

var
variable code

year
year

moncorr
A logical value indicating the use of monetary correction

Value

data frame with data for the requested variable over time

Examples

getsinimr(880, 2015)

#>      CODE         MUNICIPALITY YEAR                           VARIABLE    VALUE
#> 1   01101              IQUIQUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3074260
#> 2   01107        ALTO HOSPICIO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  6350444
#> 3   01401         POZO ALMONTE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1594943
#> 4   01402               CAMIÑA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1463976
#> 5   01403             COLCHANE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1033761
#> 6   01404                HUARA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1224376
#> 7   01405                 PICA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1096470
#> 8   02101          ANTOFAGASTA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  7233677
#> 9   02102           MEJILLONES 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1111508
#> 10  02103         SIERRA GORDA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1197074
#> 11  02104               TALTAL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1396346
#> 12  02201               CALAMA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4455655
#> 13  02202              OLLAGÜE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1252333
#> 14  02203 SAN PEDRO DE ATACAMA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2151458
#> 15  02301            TOCOPILLA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2035909
#> 16  02302          MARÍA ELENA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   808620
#> 17  03101              COPIAPÓ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  7407702
#> 18  03102              CALDERA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3701172
#> 19  03103      TIERRA AMARILLA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1222671
#> 20  03201             CHAÑARAL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3267268
#> 21  03202     DIEGO DE ALMAGRO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1244926
#> 22  03301             VALLENAR 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3503161
#> 23  03302      ALTO DEL CARMEN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1576287
#> 24  03303             FREIRINA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1272094
#> 25  03304               HUASCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1142721
#> 26  04101            LA SERENA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  8693766
#> 27  04102             COQUIMBO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 11885604
#> 28  04103            ANDACOLLO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1350837
#> 29  04104           LA HIGUERA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1069895
#> 30  04105             PAIGUANO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1673916
#> 31  04106               VICUÑA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2953653
#> 32  04201              ILLAPEL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3248934
#> 33  04202               CANELA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1542681
#> 34  04203            LOS VILOS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2929372
#> 35  04204            SALAMANCA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1978474
#> 36  04301               OVALLE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 10798360
#> 37  04302           COMBARBALÁ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2014451
#> 38  04303         MONTE PATRIA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3779588
#> 39  04304            PUNITAQUI 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1583832
#> 40  04305          RÍO HURTADO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1213042
#> 41  05101           VALPARAÍSO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  9784097
#> 42  05102           CASABLANCA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1090579
#> 43  05103               CONCÓN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1361781
#> 44  05104       JUAN FERNÁNDEZ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1168937
#> 45  05105           PUCHUNCAVÍ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1347303
#> 46  05107             QUINTERO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4702148
#> 47  05109         VIÑA DEL MAR 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  5074371
#> 48  05201       ISLA DE PASCUA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  5508316
#> 49  05301            LOS ANDES 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4737893
#> 50  05302          CALLE LARGA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1259625
#> 51  05303            RINCONADA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1075120
#> 52  05304          SAN ESTEBAN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1688999
#> 53  05401             LA LIGUA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2764363
#> 54  05402              CABILDO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1924239
#> 55  05403               PAPUDO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   674923
#> 56  05404              PETORCA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1421907
#> 57  05405             ZAPALLAR 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   858678
#> 58  05501             QUILLOTA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  5706583
#> 59  05502               CALERA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3828254
#> 60  05503             HIJUELAS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1612529
#> 61  05504              LA CRUZ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1546864
#> 62  05506              NOGALES 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1980439
#> 63  05601          SAN ANTONIO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  7009237
#> 64  05602            ALGARROBO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   782518
#> 65  05603            CARTAGENA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  5489899
#> 66  05604            EL QUISCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  5693259
#> 67  05605              EL TABO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4604252
#> 68  05606        SANTO DOMINGO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   905164
#> 69  05701           SAN FELIPE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4952106
#> 70  05702               CATEMU 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1595975
#> 71  05703             LLAILLAY 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1953574
#> 72  05704            PANQUEHUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1080263
#> 73  05705             PUTAENDO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2057132
#> 74  05706          SANTA MARÍA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1682417
#> 75  05801              QUILPUÉ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  9083692
#> 76  05802              LIMACHE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3642763
#> 77  05803                OLMUÉ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1813367
#> 78  05804        VILLA ALEMANA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  8888870
#> 79  06101             RANCAGUA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  9403120
#> 80  06102              CODEGUA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1217032
#> 81  06103               COINCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1157470
#> 82  06104             COLTAUCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1987712
#> 83  06105              DOÑIHUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2269143
#> 84  06106             GRANEROS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2291520
#> 85  06107           LAS CABRAS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1543451
#> 86  06108              MACHALÍ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1535693
#> 87  06109               MALLOA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1338908
#> 88  06110             MOSTAZAL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1801271
#> 89  06111               OLIVAR 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1149395
#> 90  06112                PEUMO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1413477
#> 91  06113           PICHIDEGUA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2072337
#> 92  06114    QUINTA DE TILCOCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1276084
#> 93  06115                RENGO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3790685
#> 94  06116             REQUINOA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1280853
#> 95  06117          SAN VICENTE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3223179
#> 96  06201            PICHILEMU 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4661982
#> 97  06202          LA ESTRELLA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   911333
#> 98  06203             LITUECHE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   985299
#> 99  06204            MARCHIHUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1181775
#> 100 06205              NAVIDAD 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3122040
#> 101 06206            PAREDONES 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1595219
#> 102 06301         SAN FERNANDO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3842607
#> 103 06302              CHÉPICA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1685584
#> 104 06303          CHIMBARONGO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2605935
#> 105 06304                LOLOL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1131897
#> 106 06305             NANCAGUA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1528424
#> 107 06306             PALMILLA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1273845
#> 108 06307            PERALILLO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1077601
#> 109 06308             PLACILLA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1222361
#> 110 06309             PUMANQUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   985935
#> 111 06310           SANTA CRUZ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2357562
#> 112 07101                TALCA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 12059145
#> 113 07102         CONSTITUCIÓN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  5989057
#> 114 07103              CUREPTO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2036270
#> 115 07104            EMPEDRADO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1512116
#> 116 07105                MAULE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2441290
#> 117 07106              PELARCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1026723
#> 118 07107             PENCAHUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1279868
#> 119 07108            RÍO CLARO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1383155
#> 120 07109         SAN CLEMENTE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3909762
#> 121 07110           SAN RAFAEL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1244098
#> 122 07201            CAUQUENES 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4526645
#> 123 07202               CHANCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1744362
#> 124 07203             PELLUHUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2502241
#> 125 07301               CURICÓ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  7727777
#> 126 07302              HUALAÑÉ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1550785
#> 127 07303             LICANTÉN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1606159
#> 128 07304               MOLINA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3704684
#> 129 07305                RAUCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1381025
#> 130 07306              ROMERAL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1103443
#> 131 07307      SAGRADA FAMILIA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1830905
#> 132 07308                 TENO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1969606
#> 133 07309            VICHUQUÉN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   916011
#> 134 07401              LINARES 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  6698395
#> 135 07402               COLBÚN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2482564
#> 136 07403              LONGAVÍ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2832441
#> 137 07404               PARRAL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3467544
#> 138 07405               RETIRO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2101113
#> 139 07406           SAN JAVIER 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3945282
#> 140 07407         VILLA ALEGRE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1911747
#> 141 07408        YERBAS BUENAS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1837550
#> 142 08101           CONCEPCIÓN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4081837
#> 143 08102              CORONEL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  8835586
#> 144 08103          CHIGUAYANTE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  7249265
#> 145 08104              FLORIDA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1743496
#> 146 08105              HUALQUI 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3062453
#> 147 08106                 LOTA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  5507398
#> 148 08107                PENCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3801694
#> 149 08108  SAN PEDRO DE LA PAZ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3755505
#> 150 08109          SANTA JUANA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2558190
#> 151 08110           TALCAHUANO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  5818035
#> 152 08111                 TOMÉ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  5884086
#> 153 08112              HUALPÉN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4588351
#> 154 08201                 LEBU 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3082844
#> 155 08202               ARAUCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3494197
#> 156 08203               CAÑETE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3675625
#> 157 08204             CONTULMO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1479577
#> 158 08205          CURANILAHUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3134671
#> 159 08206           LOS ÁLAMOS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2462886
#> 160 08207                TIRÚA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1880075
#> 161 08301          LOS ÁNGELES 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 11787152
#> 162 08302               ANTUCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1416925
#> 163 08303              CABRERO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2798941
#> 164 08304                 LAJA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2363355
#> 165 08305              MULCHÉN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2620938
#> 166 08306           NACIMIENTO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2282165
#> 167 08307              NEGRETE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1504671
#> 168 08308              QUILACO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1025032
#> 169 08309             QUILLECO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1595531
#> 170 08310          SAN ROSENDO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1180773
#> 171 08311        SANTA BÁRBARA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1747518
#> 172 08312              TUCAPEL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1925573
#> 173 08313               YUMBEL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4078460
#> 174 08314          ALTO BIOBÍO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1396018
#> 175 09101               TEMUCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 13273430
#> 176 09102              CARAHUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3556630
#> 177 09103                CUNCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2257778
#> 178 09104           CURARREHUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1493849
#> 179 09105               FREIRE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3338928
#> 180 09106            GALVARINO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2012345
#> 181 09107               GORBEA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2075606
#> 182 09108              LAUTARO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3461651
#> 183 09109             LONCOCHE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2626381
#> 184 09110            MELIPEUCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1494252
#> 185 09111       NUEVA IMPERIAL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4447584
#> 186 09112      PADRE LAS CASAS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  6791398
#> 187 09113            PERQUENCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1734570
#> 188 09114           PITRUFQUÉN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2581499
#> 189 09115                PUCÓN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1244186
#> 190 09116             SAAVEDRA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2746732
#> 191 09117      TEODORO SCHMIDT 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2433746
#> 192 09118               TOLTÉN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1849038
#> 193 09119               VILCÚN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3061660
#> 194 09120           VILLARRICA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3542918
#> 195 09121             CHOLCHOL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2060379
#> 196 09201                ANGOL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  5628638
#> 197 09202           COLLIPULLI 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2473847
#> 198 09203           CURACAUTÍN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2424757
#> 199 09204              ERCILLA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1531196
#> 200 09205            LONQUIMAY 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1915216
#> 201 09206           LOS SAUCES 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1503170
#> 202 09207               LUMACO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1909557
#> 203 09208                PURÉN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2110956
#> 204 09209              RENAICO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1507673
#> 205 09210             TRAIGUÉN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2380813
#> 206 09211             VICTORIA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3369798
#> 207 10101         PUERTO MONTT 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 10931180
#> 208 10102              CALBUCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2743014
#> 209 10103              COCHAMÓ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1147708
#> 210 10104               FRESIA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1468966
#> 211 10105            FRUTILLAR 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1478129
#> 212 10106          LOS MUERMOS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1836056
#> 213 10107           LLANQUIHUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1575890
#> 214 10108              MAULLÍN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1960491
#> 215 10109         PUERTO VARAS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1002531
#> 216 10201               CASTRO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4453910
#> 217 10202                ANCUD 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4350639
#> 218 10203              CHONCHI 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1897299
#> 219 10204      CURACO DE VÉLEZ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1372554
#> 220 10205             DALCAHUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1764069
#> 221 10206            PUQUELDÓN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1201034
#> 222 10207              QUEILÉN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1472148
#> 223 10208              QUELLÓN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2713327
#> 224 10209              QUEMCHI 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1952483
#> 225 10210             QUINCHAO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1887735
#> 226 10301               OSORNO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  7508314
#> 227 10302         PUERTO OCTAY 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1000339
#> 228 10303            PURRANQUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1921759
#> 229 10304              PUYEHUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1015329
#> 230 10305            RÍO NEGRO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1504726
#> 231 10306 SAN JUAN DE LA COSTA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2147515
#> 232 10307            SAN PABLO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1262905
#> 233 10401              CHAITÉN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1523447
#> 234 10402            FUTALEUFÚ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1256618
#> 235 10403            HUALAIHUÉ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1417463
#> 236 10404               PALENA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   937883
#> 237 11101            COIHAIQUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4253003
#> 238 11102           LAGO VERDE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   942488
#> 239 11201                AISÉN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2407488
#> 240 11202               CISNES 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1417018
#> 241 11203            GUAITECAS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1023074
#> 242 11301             COCHRANE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1118025
#> 243 11302            O´HIGGINS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   930990
#> 244 11303               TORTEL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1064802
#> 245 11401          CHILE CHICO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1791852
#> 246 11402           RÍO IBÁÑEZ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1159512
#> 247 12101         PUNTA ARENAS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4012832
#> 248 12102        LAGUNA BLANCA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   862316
#> 249 12103            RÍO VERDE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   756154
#> 250 12104         SAN GREGORIO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   786016
#> 251 12201       CABO DE HORNOS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1919260
#> 252 12301             PORVENIR 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1292217
#> 253 12302            PRIMAVERA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   832130
#> 254 12303             TIMAUKEL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   799528
#> 255 12401              NATALES 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2093886
#> 256 12402     TORRES DEL PAINE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   886576
#> 257 13101             SANTIAGO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1600755
#> 258 13102            CERRILLOS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2550083
#> 259 13103          CERRO NAVIA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 12205331
#> 260 13104             CONCHALÍ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  7076391
#> 261 13105            EL BOSQUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 14355667
#> 262 13106     ESTACIÓN CENTRAL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2023086
#> 263 13107           HUECHURABA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2214089
#> 264 13108        INDEPENDENCIA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2034598
#> 265 13109          LA CISTERNA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1771731
#> 266 13110           LA FLORIDA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 22613097
#> 267 13111            LA GRANJA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 10399881
#> 268 13112           LA PINTANA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 17835829
#> 269 13113             LA REINA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1140916
#> 270 13114           LAS CONDES 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2545452
#> 271 13115         LO BARNECHEA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1378424
#> 272 13116            LO ESPEJO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  8493987
#> 273 13117             LO PRADO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  8021835
#> 274 13118                MACUL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2582458
#> 275 13119                MAIPÚ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 40617354
#> 276 13120                ÑUÑOA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2412334
#> 277 13121  PEDRO AGUIRRE CERDA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  6027436
#> 278 13122            PEÑALOLÉN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 11181535
#> 279 13123          PROVIDENCIA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1465687
#> 280 13124             PUDAHUEL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 12017301
#> 281 13125            QUILICURA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  5778407
#> 282 13126        QUINTA NORMAL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2960731
#> 283 13127             RECOLETA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3077062
#> 284 13128                RENCA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  5131736
#> 285 13129          SAN JOAQUÍN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3180885
#> 286 13130           SAN MIGUEL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1405043
#> 287 13131            SAN RAMÓN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  7615690
#> 288 13132             VITACURA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1159430
#> 289 13201          PUENTE ALTO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 49849667
#> 290 13202               PIRQUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   844487
#> 291 13203    SAN JOSÉ DE MAIPO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1317334
#> 292 13301               COLINA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3144210
#> 293 13302                LAMPA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2369520
#> 294 13303               TILTIL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1137925
#> 295 13401         SAN BERNARDO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 14660430
#> 296 13402                 BUIN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2023952
#> 297 13403      CALERA DE TANGO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   821096
#> 298 13404                PAINE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2869600
#> 299 13501            MELIPILLA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  7420794
#> 300 13502                ALHUÉ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   950380
#> 301 13503             CURACAVÍ 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2131286
#> 302 13504          MARÍA PINTO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1177545
#> 303 13505            SAN PEDRO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1165696
#> 304 13601            TALAGANTE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4355439
#> 305 13602             EL MONTE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2189897
#> 306 13603        ISLA DE MAIPO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2357926
#> 307 13604        PADRE HURTADO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3155463
#> 308 13605             PEÑAFLOR 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  5609349
#> 309 14101             VALDIVIA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  9997817
#> 310 14102               CORRAL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1560219
#> 311 14103                LANCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2086924
#> 312 14104            LOS LAGOS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1787738
#> 313 14105                MÁFIL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1192844
#> 314 14106            MARIQUINA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1904213
#> 315 14107             PAILLACO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2051117
#> 316 14108          PANGUIPULLI 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3711286
#> 317 14201             LA UNIÓN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3344632
#> 318 14202              FUTRONO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1882667
#> 319 14203           LAGO RANCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1610361
#> 320 14204            RÍO BUENO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2671752
#> 321 15101                ARICA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 12957349
#> 322 15102            CAMARONES 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1132891
#> 323 15201                PUTRE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1603124
#> 324 15202        GENERAL LAGOS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL   955498
#> 325 16101              CHILLÁN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL 11351767
#> 326 16102               BULNES 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2234391
#> 327 16103        CHILLÁN VIEJO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2461738
#> 328 16104            EL CARMEN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1937320
#> 329 16105               PEMUCO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1388983
#> 330 16106                PINTO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1711717
#> 331 16107              QUILLÓN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  3204681
#> 332 16108          SAN IGNACIO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2112333
#> 333 16109               YUNGAY 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2038723
#> 334 16201             QUIRIHUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1942260
#> 335 16202           COBQUECURA 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1703395
#> 336 16203              COELEMU 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2386396
#> 337 16204               NINHUE 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1532461
#> 338 16205           PORTEZUELO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1348892
#> 339 16206              RÁNQUIL 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1341122
#> 340 16207             TREGUACO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1361438
#> 341 16301           SAN CARLOS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  4915689
#> 342 16302             COIHUECO 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  2832761
#> 343 16303               ÑIQUÉN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1950629
#> 344 16304           SAN FABIÁN 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1157915
#> 345 16305          SAN NICOLÁS 2015 INGRESOS POR FONDO COMÚN MUNICIPAL  1515440

getsinimr(c(880,1023), 2015)

#> Warning: data length [1035] is not a sub-multiple or multiple of the number of rows [258]

#> Error in reshape(values, idvar = c("CODE", "MUNICIPALITY"), varying = namesco(var,     year), v.names = getname(var), direction = "long", timevar = "YEAR",     times = year): length of 'varying' must be the product of length of 'v.names' and length of 'times'