Search for a SINIM variable and if no keyword is provided returns all variables

Search for a SINIM variable and if no keyword is provided returns all variables

search_sinim_vars(keyword)

Arguments

keyword

keyword

Value

data frame with results

Examples

search_sinim_vars("ingresos propios")
#>     code
#> 1    879
#> 3    882
#> 6    890
#> 9   1262
#> 38  1149
#> 48  4339
#> 66  3771
#> 73  1086
#> 76  1089
#> 79  1272
#> 147 1138
#> 265 1140
#>                                                                                                          variable
#> 1                                                                              Ingresos Propios Permanentes (IPP)
#> 3                                                                                    Ingresos Propios (IPP y FCM)
#> 6                                         Ingresos Propios, según criterio de Contraloría General de la República
#> 9                                                                  Ingresos Propios Permanentes per Cápita (IPPP)
#> 38  Participación Gasto en Personal sobre los Ingresos Propios (límite del 35%) criterio CGR válido hasta el 2016
#> 48               Umbral Legal (42%) de Gastos en Personal (en miles de pesos) Ley 20.922 (vigente desde año 2016)
#> 66               Participación de Gastos en Bienes y Servicios de Consumo sobre los Ingresos Propios (desde 2008)
#> 73                                        Participación de Ingresos Propios Permanentes (IPP) en el Ingreso Total
#> 76          Participación de Ingresos Propios Permanentes sobre el Ingreso Total (descontadas las transferencias)
#> 79                                               Dependencia del Fondo Común Municipal sobre los Ingresos Propios
#> 147                                         Transferencias Municipales a Salud sobre Ingresos Propios Municipales
#> 265                                     Transferencias Municipales a Educación sobre Ingresos Propios Municipales
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          description
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Indicador Bep Ingresos (variable sistema antiguo)
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Indicador Bep Ingresos (variable sistema antiguo)
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Indicador Bep Ingresos (variable sistema antiguo)
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Ingresos Propios Permanentes per Cápita.
#> 38  Identificar que el total de gasto en personal, no supere el máximo legal del 35% del rendimiento estimado de los ingresos señalados en el art. 1 de la Ley 18.294, modificada por la letra a) del Art. 67 de la Ley 18.382.  NOTA:  Este indicador considera la Renta del Alcalde(sa), ya que a la fecha ésta remuneración está considerada en el subtítulo 21, y no es posible extraer el monto específico de la renta.    Además, este indicador no considera el aporte a los servicios de bienestar estipulado en la Ley 19.754, Art. 3.   Por otro lado, a partir del año 2008, aquellas municipalidades que se excedan según lo dispuesto en la Ley 20.198, no estarán obligadas a ajustarse a dicha restricción (Art. 5°).
#> 48                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            Umbral Legal (42%) de Gastos en Personal (en miles de pesos) Ley 20.922. Para análisis y proyección se debe observar y calcular con el año anterior, según lo indicado como Ingresos Propios Percibidos criterio CGR. 
#> 66                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              Indica el porcentaje de gastos de consumo interno y servicios hacia la comunidad como alumbrado público y servicios de aseo, sobre los Ingresos Propios (IPP   FCM).
#> 73                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           A mayor valor del indicador, mayor autonomía financiera del Municipio. Esto es, mayor proporción de los ingresos del presupuesto provienen de fuentes propias y de carácter permanente.
#> 76                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 A mayor valor del indicador, mayor autonomía financiera del Municipio. Esto significa una mayor proporción de los ingresos del presupuesto provienen de fuentes propias y de carácter permanente.
#> 79                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          Mide la dependencia del Fondo Común Municipal sobre los Ingresos Propios
#> 147                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   Aportes municipales al sector de salud sobre el Ingreso Propio (IP) municipal.
#> 265                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              Transferencias de la municipalidad al sector Educación sobre los Ingresos Propios (IP) municipales.
#>                                           area
#> 1   01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 3   01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 6   01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 9   01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 38  01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 48  01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 66  01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 73  01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 76  01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 79  01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 147                       04.  SALUD MUNICIPAL
#> 265                   03.  EDUCACION MUNICIPAL
#>                                subarea unit
#> 1         A. INGRESOS MUNICIPALES (M$) M$  
#> 3         A. INGRESOS MUNICIPALES (M$) M$  
#> 6         A. INGRESOS MUNICIPALES (M$) M$  
#> 9         A. INGRESOS MUNICIPALES (M$) M$  
#> 38               E. GASTOS EN PERSONAL %   
#> 48               E. GASTOS EN PERSONAL M$  
#> 66           D. GASTOS MUNICIPALES (%) %   
#> 73         B. INGRESOS MUNICIPALES (%) %   
#> 76         B. INGRESOS MUNICIPALES (%) %   
#> 79         B. INGRESOS MUNICIPALES (%) %   
#> 147    C. INGRESOS  EN SALUD MUNICIPAL %   
#> 265 D. INGRESOS EN EDUCACION MUNICIPAL %