Search a SINIM variable

Search a SINIM variable

searchsinimvar(keyword)

Arguments

keyword
keyword

Value

data frame with results

Examples

searchsinimvar("ingresos propios")

#>     CODE
#> 1    879
#> 3    882
#> 6    890
#> 9   1262
#> 38  1149
#> 67  3771
#> 74  1086
#> 77  1089
#> 80  1272
#> 135 1138
#> 229 1140
#>                                                                                                          VARIABLE
#> 1                                                                              Ingresos Propios Permanentes (IPP)
#> 3                                                                                    Ingresos Propios (IPP y FCM)
#> 6                                         Ingresos Propios, según criterio de Contraloría General de la República
#> 9                                                                  Ingresos Propios Permanentes per Cápita (IPPP)
#> 38  Participación Gasto en Personal sobre los Ingresos Propios (límite del 35%) criterio CGR válido hasta el 2016
#> 67               Participación de Gastos en Bienes y Servicios de Consumo sobre los Ingresos Propios (desde 2008)
#> 74                                        Participación de Ingresos Propios Permanentes (IPP) en el Ingreso Total
#> 77          Participación de Ingresos Propios Permanentes sobre el Ingreso Total (descontadas las transferencias)
#> 80                                               Dependencia del Fondo Común Municipal sobre los Ingresos Propios
#> 135                                         Transferencias Municipales a Salud sobre Ingresos Propios Municipales
#> 229                                     Transferencias Municipales a Educación sobre Ingresos Propios Municipales
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          DESCRIPTION
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Indicador Bep Ingresos (variable sistema antiguo)
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Indicador Bep Ingresos (variable sistema antiguo)
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Indicador Bep Ingresos (variable sistema antiguo)
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Ingresos Propios Permanentes per Cápita.
#> 38  Identificar que el total de gasto en personal, no supere el máximo legal del 35% del rendimiento estimado de los ingresos señalados en el art. 1 de la Ley 18.294, modificada por la letra a) del Art. 67 de la Ley 18.382.  NOTA:  Este indicador considera la Renta del Alcalde(sa), ya que a la fecha ésta remuneración está considerada en el subtítulo 21, y no es posible extraer el monto específico de la renta.    Además, este indicador no considera el aporte a los servicios de bienestar estipulado en la Ley 19.754, Art. 3.   Por otro lado, a partir del año 2008, aquellas municipalidades que se excedan según lo dispuesto en la Ley 20.198, no estarán obligadas a ajustarse a dicha restricción (Art. 5°).
#> 67                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              Indica el porcentaje de gastos de consumo interno y servicios hacia la comunidad como alumbrado público y servicios de aseo, sobre los Ingresos Propios (IPP   FCM).
#> 74                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           A mayor valor del indicador, mayor autonomía financiera del Municipio. Esto es, mayor proporción de los ingresos del presupuesto provienen de fuentes propias y de carácter permanente.
#> 77                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 A mayor valor del indicador, mayor autonomía financiera del Municipio. Esto significa una mayor proporción de los ingresos del presupuesto provienen de fuentes propias y de carácter permanente.
#> 80                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       Mide la dependencia de los Ingresos Propios sobre el Fondo Común Municipal.
#> 135                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   Aportes municipales al sector de salud sobre el Ingreso Propio (IP) municipal.
#> 229                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              Transferencias de la municipalidad al sector Educación sobre los Ingresos Propios (IP) municipales.
#>                                           AREA
#> 1   01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 3   01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 6   01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 9   01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 38  01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 67  01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 74  01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 77  01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 80  01.  ADMINISTRACION Y FINANZAS MUNICIPALES
#> 135                       04.  SALUD MUNICIPAL
#> 229                   03.  EDUCACION MUNICIPAL
#>                                SUBAREA UNIT
#> 1         A. INGRESOS MUNICIPALES (M$) M$  
#> 3         A. INGRESOS MUNICIPALES (M$) M$  
#> 6         A. INGRESOS MUNICIPALES (M$) M$  
#> 9         A. INGRESOS MUNICIPALES (M$) M$  
#> 38               E. GASTOS EN PERSONAL %   
#> 67           D. GASTOS MUNICIPALES (%) %   
#> 74         B. INGRESOS MUNICIPALES (%) %   
#> 77         B. INGRESOS MUNICIPALES (%) %   
#> 80         B. INGRESOS MUNICIPALES (%) %   
#> 135    C. INGRESOS  EN SALUD MUNICIPAL %   
#> 229 D. INGRESOS EN EDUCACION MUNICIPAL %