Get a SINIM variable data in a specific year as a data frame

Get a SINIM variable data in a specific year as a data frame

getsinimr(var, year, moncorr = T)

Arguments

var
variable code

year
year

moncorr
A logical value indicating the use of monetary correction

Value

data frame with data for the requested variable over time

Examples

getsinimr(880, 2015)

#>      CODE         MUNICIPALITY INGRESOS POR FONDO COMÚN MUNICIPAL
#> 1   01101              IQUIQUE                            3074260
#> 2   01107        ALTO HOSPICIO                            6350444
#> 3   01401         POZO ALMONTE                            1594943
#> 4   01402               CAMIÑA                            1463976
#> 5   01403             COLCHANE                            1033761
#> 6   01404                HUARA                            1224376
#> 7   01405                 PICA                            1096470
#> 8   02101          ANTOFAGASTA                            7233677
#> 9   02102           MEJILLONES                            1111508
#> 10  02103         SIERRA GORDA                            1197074
#> 11  02104               TALTAL                            1396346
#> 12  02201               CALAMA                            4455655
#> 13  02202              OLLAGÜE                            1252333
#> 14  02203 SAN PEDRO DE ATACAMA                            2151458
#> 15  02301            TOCOPILLA                            2035909
#> 16  02302          MARÍA ELENA                             808620
#> 17  03101              COPIAPÓ                            7407702
#> 18  03102              CALDERA                            3701172
#> 19  03103      TIERRA AMARILLA                            1222671
#> 20  03201             CHAÑARAL                            3267268
#> 21  03202     DIEGO DE ALMAGRO                            1244926
#> 22  03301             VALLENAR                            3503161
#> 23  03302      ALTO DEL CARMEN                            1576287
#> 24  03303             FREIRINA                            1272094
#> 25  03304               HUASCO                            1142721
#> 26  04101            LA SERENA                            8693766
#> 27  04102             COQUIMBO                           11885604
#> 28  04103            ANDACOLLO                            1350837
#> 29  04104           LA HIGUERA                            1069895
#> 30  04105             PAIGUANO                            1673916
#> 31  04106               VICUÑA                            2953653
#> 32  04201              ILLAPEL                            3248934
#> 33  04202               CANELA                            1542681
#> 34  04203            LOS VILOS                            2929372
#> 35  04204            SALAMANCA                            1978474
#> 36  04301               OVALLE                           10798360
#> 37  04302           COMBARBALÁ                            2014451
#> 38  04303         MONTE PATRIA                            3779588
#> 39  04304            PUNITAQUI                            1583832
#> 40  04305          RÍO HURTADO                            1213042
#> 41  05101           VALPARAÍSO                            9784097
#> 42  05102           CASABLANCA                            1090579
#> 43  05103               CONCÓN                            1361781
#> 44  05104       JUAN FERNÁNDEZ                            1168937
#> 45  05105           PUCHUNCAVÍ                            1347303
#> 46  05107             QUINTERO                            4702148
#> 47  05109         VIÑA DEL MAR                            5074371
#> 48  05201       ISLA DE PASCUA                            5508316
#> 49  05301            LOS ANDES                            4737893
#> 50  05302          CALLE LARGA                            1259625
#> 51  05303            RINCONADA                            1075120
#> 52  05304          SAN ESTEBAN                            1688999
#> 53  05401             LA LIGUA                            2764363
#> 54  05402              CABILDO                            1924239
#> 55  05403               PAPUDO                             674923
#> 56  05404              PETORCA                            1421907
#> 57  05405             ZAPALLAR                             858678
#> 58  05501             QUILLOTA                            5706583
#> 59  05502               CALERA                            3828254
#> 60  05503             HIJUELAS                            1612529
#> 61  05504              LA CRUZ                            1546864
#> 62  05506              NOGALES                            1980439
#> 63  05601          SAN ANTONIO                            7009237
#> 64  05602            ALGARROBO                             782518
#> 65  05603            CARTAGENA                            5489899
#> 66  05604            EL QUISCO                            5693259
#> 67  05605              EL TABO                            4604252
#> 68  05606        SANTO DOMINGO                             905164
#> 69  05701           SAN FELIPE                            4952106
#> 70  05702               CATEMU                            1595975
#> 71  05703             LLAILLAY                            1953574
#> 72  05704            PANQUEHUE                            1080263
#> 73  05705             PUTAENDO                            2057132
#> 74  05706          SANTA MARÍA                            1682417
#> 75  05801              QUILPUÉ                            9083692
#> 76  05802              LIMACHE                            3642763
#> 77  05803                OLMUÉ                            1813367
#> 78  05804        VILLA ALEMANA                            8888870
#> 79  06101             RANCAGUA                            9403120
#> 80  06102              CODEGUA                            1217032
#> 81  06103               COINCO                            1157470
#> 82  06104             COLTAUCO                            1987712
#> 83  06105              DOÑIHUE                            2269143
#> 84  06106             GRANEROS                            2291520
#> 85  06107           LAS CABRAS                            1543451
#> 86  06108              MACHALÍ                            1535693
#> 87  06109               MALLOA                            1338908
#> 88  06110             MOSTAZAL                            1801271
#> 89  06111               OLIVAR                            1149395
#> 90  06112                PEUMO                            1413477
#> 91  06113           PICHIDEGUA                            2072337
#> 92  06114    QUINTA DE TILCOCO                            1276084
#> 93  06115                RENGO                            3790685
#> 94  06116             REQUINOA                            1280853
#> 95  06117          SAN VICENTE                            3223179
#> 96  06201            PICHILEMU                            4661982
#> 97  06202          LA ESTRELLA                             911333
#> 98  06203             LITUECHE                             985299
#> 99  06204            MARCHIHUE                            1181775
#> 100 06205              NAVIDAD                            3122040
#> 101 06206            PAREDONES                            1595219
#> 102 06301         SAN FERNANDO                            3842607
#> 103 06302              CHÉPICA                            1685584
#> 104 06303          CHIMBARONGO                            2605935
#> 105 06304                LOLOL                            1131897
#> 106 06305             NANCAGUA                            1528424
#> 107 06306             PALMILLA                            1273845
#> 108 06307            PERALILLO                            1077601
#> 109 06308             PLACILLA                            1222361
#> 110 06309             PUMANQUE                             985935
#> 111 06310           SANTA CRUZ                            2357562
#> 112 07101                TALCA                           12059145
#> 113 07102         CONSTITUCIÓN                            5989057
#> 114 07103              CUREPTO                            2036270
#> 115 07104            EMPEDRADO                            1512116
#> 116 07105                MAULE                            2441290
#> 117 07106              PELARCO                            1026723
#> 118 07107             PENCAHUE                            1279868
#> 119 07108            RÍO CLARO                            1383155
#> 120 07109         SAN CLEMENTE                            3909762
#> 121 07110           SAN RAFAEL                            1244098
#> 122 07201            CAUQUENES                            4526645
#> 123 07202               CHANCO                            1744362
#> 124 07203             PELLUHUE                            2502241
#> 125 07301               CURICÓ                            7727777
#> 126 07302              HUALAÑÉ                            1550785
#> 127 07303             LICANTÉN                            1606159
#> 128 07304               MOLINA                            3704684
#> 129 07305                RAUCO                            1381025
#> 130 07306              ROMERAL                            1103443
#> 131 07307      SAGRADA FAMILIA                            1830905
#> 132 07308                 TENO                            1969606
#> 133 07309            VICHUQUÉN                             916011
#> 134 07401              LINARES                            6698395
#> 135 07402               COLBÚN                            2482564
#> 136 07403              LONGAVÍ                            2832441
#> 137 07404               PARRAL                            3467544
#> 138 07405               RETIRO                            2101113
#> 139 07406           SAN JAVIER                            3945282
#> 140 07407         VILLA ALEGRE                            1911747
#> 141 07408        YERBAS BUENAS                            1837550
#> 142 08101           CONCEPCIÓN                            4081837
#> 143 08102              CORONEL                            8835586
#> 144 08103          CHIGUAYANTE                            7249265
#> 145 08104              FLORIDA                            1743496
#> 146 08105              HUALQUI                            3062453
#> 147 08106                 LOTA                            5507398
#> 148 08107                PENCO                            3801694
#> 149 08108  SAN PEDRO DE LA PAZ                            3755505
#> 150 08109          SANTA JUANA                            2558190
#> 151 08110           TALCAHUANO                            5818035
#> 152 08111                 TOMÉ                            5884086
#> 153 08112              HUALPÉN                            4588351
#> 154 08201                 LEBU                            3082844
#> 155 08202               ARAUCO                            3494197
#> 156 08203               CAÑETE                            3675625
#> 157 08204             CONTULMO                            1479577
#> 158 08205          CURANILAHUE                            3134671
#> 159 08206           LOS ÁLAMOS                            2462886
#> 160 08207                TIRÚA                            1880075
#> 161 08301          LOS ÁNGELES                           11787152
#> 162 08302               ANTUCO                            1416925
#> 163 08303              CABRERO                            2798941
#> 164 08304                 LAJA                            2363355
#> 165 08305              MULCHÉN                            2620938
#> 166 08306           NACIMIENTO                            2282165
#> 167 08307              NEGRETE                            1504671
#> 168 08308              QUILACO                            1025032
#> 169 08309             QUILLECO                            1595531
#> 170 08310          SAN ROSENDO                            1180773
#> 171 08311        SANTA BÁRBARA                            1747518
#> 172 08312              TUCAPEL                            1925573
#> 173 08313               YUMBEL                            4078460
#> 174 08314          ALTO BIOBÍO                            1396018
#> 175 09101               TEMUCO                           13273430
#> 176 09102              CARAHUE                            3556630
#> 177 09103                CUNCO                            2257778
#> 178 09104           CURARREHUE                            1493849
#> 179 09105               FREIRE                            3338928
#> 180 09106            GALVARINO                            2012345
#> 181 09107               GORBEA                            2075606
#> 182 09108              LAUTARO                            3461651
#> 183 09109             LONCOCHE                            2626381
#> 184 09110            MELIPEUCO                            1494252
#> 185 09111       NUEVA IMPERIAL                            4447584
#> 186 09112      PADRE LAS CASAS                            6791398
#> 187 09113            PERQUENCO                            1734570
#> 188 09114           PITRUFQUÉN                            2581499
#> 189 09115                PUCÓN                            1244186
#> 190 09116             SAAVEDRA                            2746732
#> 191 09117      TEODORO SCHMIDT                            2433746
#> 192 09118               TOLTÉN                            1849038
#> 193 09119               VILCÚN                            3061660
#> 194 09120           VILLARRICA                            3542918
#> 195 09121             CHOLCHOL                            2060379
#> 196 09201                ANGOL                            5628638
#> 197 09202           COLLIPULLI                            2473847
#> 198 09203           CURACAUTÍN                            2424757
#> 199 09204              ERCILLA                            1531196
#> 200 09205            LONQUIMAY                            1915216
#> 201 09206           LOS SAUCES                            1503170
#> 202 09207               LUMACO                            1909557
#> 203 09208                PURÉN                            2110956
#> 204 09209              RENAICO                            1507673
#> 205 09210             TRAIGUÉN                            2380813
#> 206 09211             VICTORIA                            3369798
#> 207 10101         PUERTO MONTT                           10931180
#> 208 10102              CALBUCO                            2743014
#> 209 10103              COCHAMÓ                            1147708
#> 210 10104               FRESIA                            1468966
#> 211 10105            FRUTILLAR                            1478129
#> 212 10106          LOS MUERMOS                            1836056
#> 213 10107           LLANQUIHUE                            1575890
#> 214 10108              MAULLÍN                            1960491
#> 215 10109         PUERTO VARAS                            1002531
#> 216 10201               CASTRO                            4453910
#> 217 10202                ANCUD                            4350639
#> 218 10203              CHONCHI                            1897299
#> 219 10204      CURACO DE VÉLEZ                            1372554
#> 220 10205             DALCAHUE                            1764069
#> 221 10206            PUQUELDÓN                            1201034
#> 222 10207              QUEILÉN                            1472148
#> 223 10208              QUELLÓN                            2713327
#> 224 10209              QUEMCHI                            1952483
#> 225 10210             QUINCHAO                            1887735
#> 226 10301               OSORNO                            7508314
#> 227 10302         PUERTO OCTAY                            1000339
#> 228 10303            PURRANQUE                            1921759
#> 229 10304              PUYEHUE                            1015329
#> 230 10305            RÍO NEGRO                            1504726
#> 231 10306 SAN JUAN DE LA COSTA                            2147515
#> 232 10307            SAN PABLO                            1262905
#> 233 10401              CHAITÉN                            1523447
#> 234 10402            FUTALEUFÚ                            1256618
#> 235 10403            HUALAIHUÉ                            1417463
#> 236 10404               PALENA                             937883
#> 237 11101            COIHAIQUE                            4253003
#> 238 11102           LAGO VERDE                             942488
#> 239 11201                AISÉN                            2407488
#> 240 11202               CISNES                            1417018
#> 241 11203            GUAITECAS                            1023074
#> 242 11301             COCHRANE                            1118025
#> 243 11302            O´HIGGINS                             930990
#> 244 11303               TORTEL                            1064802
#> 245 11401          CHILE CHICO                            1791852
#> 246 11402           RÍO IBÁÑEZ                            1159512
#> 247 12101         PUNTA ARENAS                            4012832
#> 248 12102        LAGUNA BLANCA                             862316
#> 249 12103            RÍO VERDE                             756154
#> 250 12104         SAN GREGORIO                             786016
#> 251 12201       CABO DE HORNOS                            1919260
#> 252 12301             PORVENIR                            1292217
#> 253 12302            PRIMAVERA                             832130
#> 254 12303             TIMAUKEL                             799528
#> 255 12401              NATALES                            2093886
#> 256 12402     TORRES DEL PAINE                             886576
#> 257 13101             SANTIAGO                            1600755
#> 258 13102            CERRILLOS                            2550083
#> 259 13103          CERRO NAVIA                           12205331
#> 260 13104             CONCHALÍ                            7076391
#> 261 13105            EL BOSQUE                           14355667
#> 262 13106     ESTACIÓN CENTRAL                            2023086
#> 263 13107           HUECHURABA                            2214089
#> 264 13108        INDEPENDENCIA                            2034598
#> 265 13109          LA CISTERNA                            1771731
#> 266 13110           LA FLORIDA                           22613097
#> 267 13111            LA GRANJA                           10399881
#> 268 13112           LA PINTANA                           17835829
#> 269 13113             LA REINA                            1140916
#> 270 13114           LAS CONDES                            2545452
#> 271 13115         LO BARNECHEA                            1378424
#> 272 13116            LO ESPEJO                            8493987
#> 273 13117             LO PRADO                            8021835
#> 274 13118                MACUL                            2582458
#> 275 13119                MAIPÚ                           40617354
#> 276 13120                ÑUÑOA                            2412334
#> 277 13121  PEDRO AGUIRRE CERDA                            6027436
#> 278 13122            PEÑALOLÉN                           11181535
#> 279 13123          PROVIDENCIA                            1465687
#> 280 13124             PUDAHUEL                           12017301
#> 281 13125            QUILICURA                            5778407
#> 282 13126        QUINTA NORMAL                            2960731
#> 283 13127             RECOLETA                            3077062
#> 284 13128                RENCA                            5131736
#> 285 13129          SAN JOAQUÍN                            3180885
#> 286 13130           SAN MIGUEL                            1405043
#> 287 13131            SAN RAMÓN                            7615690
#> 288 13132             VITACURA                            1159430
#> 289 13201          PUENTE ALTO                           49849667
#> 290 13202               PIRQUE                             844487
#> 291 13203    SAN JOSÉ DE MAIPO                            1317334
#> 292 13301               COLINA                            3144210
#> 293 13302                LAMPA                            2369520
#> 294 13303               TILTIL                            1137925
#> 295 13401         SAN BERNARDO                           14660430
#> 296 13402                 BUIN                            2023952
#> 297 13403      CALERA DE TANGO                             821096
#> 298 13404                PAINE                            2869600
#> 299 13501            MELIPILLA                            7420794
#> 300 13502                ALHUÉ                             950380
#> 301 13503             CURACAVÍ                            2131286
#> 302 13504          MARÍA PINTO                            1177545
#> 303 13505            SAN PEDRO                            1165696
#> 304 13601            TALAGANTE                            4355439
#> 305 13602             EL MONTE                            2189897
#> 306 13603        ISLA DE MAIPO                            2357926
#> 307 13604        PADRE HURTADO                            3155463
#> 308 13605             PEÑAFLOR                            5609349
#> 309 14101             VALDIVIA                            9997817
#> 310 14102               CORRAL                            1560219
#> 311 14103                LANCO                            2086924
#> 312 14104            LOS LAGOS                            1787738
#> 313 14105                MÁFIL                            1192844
#> 314 14106            MARIQUINA                            1904213
#> 315 14107             PAILLACO                            2051117
#> 316 14108          PANGUIPULLI                            3711286
#> 317 14201             LA UNIÓN                            3344632
#> 318 14202              FUTRONO                            1882667
#> 319 14203           LAGO RANCO                            1610361
#> 320 14204            RÍO BUENO                            2671752
#> 321 15101                ARICA                           12957349
#> 322 15102            CAMARONES                            1132891
#> 323 15201                PUTRE                            1603124
#> 324 15202        GENERAL LAGOS                             955498
#> 325 16101              CHILLÁN                           11351767
#> 326 16102               BULNES                            2234391
#> 327 16103        CHILLÁN VIEJO                            2461738
#> 328 16104            EL CARMEN                            1937320
#> 329 16105               PEMUCO                            1388983
#> 330 16106                PINTO                            1711717
#> 331 16107              QUILLÓN                            3204681
#> 332 16108          SAN IGNACIO                            2112333
#> 333 16109               YUNGAY                            2038723
#> 334 16201             QUIRIHUE                            1942260
#> 335 16202           COBQUECURA                            1703395
#> 336 16203              COELEMU                            2386396
#> 337 16204               NINHUE                            1532461
#> 338 16205           PORTEZUELO                            1348892
#> 339 16206              RÁNQUIL                            1341122
#> 340 16207             TREGUACO                            1361438
#> 341 16301           SAN CARLOS                            4915689
#> 342 16302             COIHUECO                            2832761
#> 343 16303               ÑIQUÉN                            1950629
#> 344 16304           SAN FABIÁN                            1157915
#> 345 16305          SAN NICOLÁS                            1515440

getsinimr(c(880,1023), 2015)

#>      CODE         MUNICIPALITY INGRESOS POR FONDO COMÚN MUNICIPAL
#> 1   01101              IQUIQUE                            3074260
#> 2   01107        ALTO HOSPICIO                            6350444
#> 3   01401         POZO ALMONTE                            1594943
#> 4   01402               CAMIÑA                            1463976
#> 5   01403             COLCHANE                            1033761
#> 6   01404                HUARA                            1224376
#> 7   01405                 PICA                            1096470
#> 8   02101          ANTOFAGASTA                            7233677
#> 9   02102           MEJILLONES                            1111508
#> 10  02103         SIERRA GORDA                            1197074
#> 11  02104               TALTAL                            1396346
#> 12  02201               CALAMA                            4455655
#> 13  02202              OLLAGÜE                            1252333
#> 14  02203 SAN PEDRO DE ATACAMA                            2151458
#> 15  02301            TOCOPILLA                            2035909
#> 16  02302          MARÍA ELENA                             808620
#> 17  03101              COPIAPÓ                            7407702
#> 18  03102              CALDERA                            3701172
#> 19  03103      TIERRA AMARILLA                            1222671
#> 20  03201             CHAÑARAL                            3267268
#> 21  03202     DIEGO DE ALMAGRO                            1244926
#> 22  03301             VALLENAR                            3503161
#> 23  03302      ALTO DEL CARMEN                            1576287
#> 24  03303             FREIRINA                            1272094
#> 25  03304               HUASCO                            1142721
#> 26  04101            LA SERENA                            8693766
#> 27  04102             COQUIMBO                           11885604
#> 28  04103            ANDACOLLO                            1350837
#> 29  04104           LA HIGUERA                            1069895
#> 30  04105             PAIGUANO                            1673916
#> 31  04106               VICUÑA                            2953653
#> 32  04201              ILLAPEL                            3248934
#> 33  04202               CANELA                            1542681
#> 34  04203            LOS VILOS                            2929372
#> 35  04204            SALAMANCA                            1978474
#> 36  04301               OVALLE                           10798360
#> 37  04302           COMBARBALÁ                            2014451
#> 38  04303         MONTE PATRIA                            3779588
#> 39  04304            PUNITAQUI                            1583832
#> 40  04305          RÍO HURTADO                            1213042
#> 41  05101           VALPARAÍSO                            9784097
#> 42  05102           CASABLANCA                            1090579
#> 43  05103               CONCÓN                            1361781
#> 44  05104       JUAN FERNÁNDEZ                            1168937
#> 45  05105           PUCHUNCAVÍ                            1347303
#> 46  05107             QUINTERO                            4702148
#> 47  05109         VIÑA DEL MAR                            5074371
#> 48  05201       ISLA DE PASCUA                            5508316
#> 49  05301            LOS ANDES                            4737893
#> 50  05302          CALLE LARGA                            1259625
#> 51  05303            RINCONADA                            1075120
#> 52  05304          SAN ESTEBAN                            1688999
#> 53  05401             LA LIGUA                            2764363
#> 54  05402              CABILDO                            1924239
#> 55  05403               PAPUDO                             674923
#> 56  05404              PETORCA                            1421907
#> 57  05405             ZAPALLAR                             858678
#> 58  05501             QUILLOTA                            5706583
#> 59  05502               CALERA                            3828254
#> 60  05503             HIJUELAS                            1612529
#> 61  05504              LA CRUZ                            1546864
#> 62  05506              NOGALES                            1980439
#> 63  05601          SAN ANTONIO                            7009237
#> 64  05602            ALGARROBO                             782518
#> 65  05603            CARTAGENA                            5489899
#> 66  05604            EL QUISCO                            5693259
#> 67  05605              EL TABO                            4604252
#> 68  05606        SANTO DOMINGO                             905164
#> 69  05701           SAN FELIPE                            4952106
#> 70  05702               CATEMU                            1595975
#> 71  05703             LLAILLAY                            1953574
#> 72  05704            PANQUEHUE                            1080263
#> 73  05705             PUTAENDO                            2057132
#> 74  05706          SANTA MARÍA                            1682417
#> 75  05801              QUILPUÉ                            9083692
#> 76  05802              LIMACHE                            3642763
#> 77  05803                OLMUÉ                            1813367
#> 78  05804        VILLA ALEMANA                            8888870
#> 79  06101             RANCAGUA                            9403120
#> 80  06102              CODEGUA                            1217032
#> 81  06103               COINCO                            1157470
#> 82  06104             COLTAUCO                            1987712
#> 83  06105              DOÑIHUE                            2269143
#> 84  06106             GRANEROS                            2291520
#> 85  06107           LAS CABRAS                            1543451
#> 86  06108              MACHALÍ                            1535693
#> 87  06109               MALLOA                            1338908
#> 88  06110             MOSTAZAL                            1801271
#> 89  06111               OLIVAR                            1149395
#> 90  06112                PEUMO                            1413477
#> 91  06113           PICHIDEGUA                            2072337
#> 92  06114    QUINTA DE TILCOCO                            1276084
#> 93  06115                RENGO                            3790685
#> 94  06116             REQUINOA                            1280853
#> 95  06117          SAN VICENTE                            3223179
#> 96  06201            PICHILEMU                            4661982
#> 97  06202          LA ESTRELLA                             911333
#> 98  06203             LITUECHE                             985299
#> 99  06204            MARCHIHUE                            1181775
#> 100 06205              NAVIDAD                            3122040
#> 101 06206            PAREDONES                            1595219
#> 102 06301         SAN FERNANDO                            3842607
#> 103 06302              CHÉPICA                            1685584
#> 104 06303          CHIMBARONGO                            2605935
#> 105 06304                LOLOL                            1131897
#> 106 06305             NANCAGUA                            1528424
#> 107 06306             PALMILLA                            1273845
#> 108 06307            PERALILLO                            1077601
#> 109 06308             PLACILLA                            1222361
#> 110 06309             PUMANQUE                             985935
#> 111 06310           SANTA CRUZ                            2357562
#> 112 07101                TALCA                           12059145
#> 113 07102         CONSTITUCIÓN                            5989057
#> 114 07103              CUREPTO                            2036270
#> 115 07104            EMPEDRADO                            1512116
#> 116 07105                MAULE                            2441290
#> 117 07106              PELARCO                            1026723
#> 118 07107             PENCAHUE                            1279868
#> 119 07108            RÍO CLARO                            1383155
#> 120 07109         SAN CLEMENTE                            3909762
#> 121 07110           SAN RAFAEL                            1244098
#> 122 07201            CAUQUENES                            4526645
#> 123 07202               CHANCO                            1744362
#> 124 07203             PELLUHUE                            2502241
#> 125 07301               CURICÓ                            7727777
#> 126 07302              HUALAÑÉ                            1550785
#> 127 07303             LICANTÉN                            1606159
#> 128 07304               MOLINA                            3704684
#> 129 07305                RAUCO                            1381025
#> 130 07306              ROMERAL                            1103443
#> 131 07307      SAGRADA FAMILIA                            1830905
#> 132 07308                 TENO                            1969606
#> 133 07309            VICHUQUÉN                             916011
#> 134 07401              LINARES                            6698395
#> 135 07402               COLBÚN                            2482564
#> 136 07403              LONGAVÍ                            2832441
#> 137 07404               PARRAL                            3467544
#> 138 07405               RETIRO                            2101113
#> 139 07406           SAN JAVIER                            3945282
#> 140 07407         VILLA ALEGRE                            1911747
#> 141 07408        YERBAS BUENAS                            1837550
#> 142 08101           CONCEPCIÓN                            4081837
#> 143 08102              CORONEL                            8835586
#> 144 08103          CHIGUAYANTE                            7249265
#> 145 08104              FLORIDA                            1743496
#> 146 08105              HUALQUI                            3062453
#> 147 08106                 LOTA                            5507398
#> 148 08107                PENCO                            3801694
#> 149 08108  SAN PEDRO DE LA PAZ                            3755505
#> 150 08109          SANTA JUANA                            2558190
#> 151 08110           TALCAHUANO                            5818035
#> 152 08111                 TOMÉ                            5884086
#> 153 08112              HUALPÉN                            4588351
#> 154 08201                 LEBU                            3082844
#> 155 08202               ARAUCO                            3494197
#> 156 08203               CAÑETE                            3675625
#> 157 08204             CONTULMO                            1479577
#> 158 08205          CURANILAHUE                            3134671
#> 159 08206           LOS ÁLAMOS                            2462886
#> 160 08207                TIRÚA                            1880075
#> 161 08301          LOS ÁNGELES                           11787152
#> 162 08302               ANTUCO                            1416925
#> 163 08303              CABRERO                            2798941
#> 164 08304                 LAJA                            2363355
#> 165 08305              MULCHÉN                            2620938
#> 166 08306           NACIMIENTO                            2282165
#> 167 08307              NEGRETE                            1504671
#> 168 08308              QUILACO                            1025032
#> 169 08309             QUILLECO                            1595531
#> 170 08310          SAN ROSENDO                            1180773
#> 171 08311        SANTA BÁRBARA                            1747518
#> 172 08312              TUCAPEL                            1925573
#> 173 08313               YUMBEL                            4078460
#> 174 08314          ALTO BIOBÍO                            1396018
#> 175 09101               TEMUCO                           13273430
#> 176 09102              CARAHUE                            3556630
#> 177 09103                CUNCO                            2257778
#> 178 09104           CURARREHUE                            1493849
#> 179 09105               FREIRE                            3338928
#> 180 09106            GALVARINO                            2012345
#> 181 09107               GORBEA                            2075606
#> 182 09108              LAUTARO                            3461651
#> 183 09109             LONCOCHE                            2626381
#> 184 09110            MELIPEUCO                            1494252
#> 185 09111       NUEVA IMPERIAL                            4447584
#> 186 09112      PADRE LAS CASAS                            6791398
#> 187 09113            PERQUENCO                            1734570
#> 188 09114           PITRUFQUÉN                            2581499
#> 189 09115                PUCÓN                            1244186
#> 190 09116             SAAVEDRA                            2746732
#> 191 09117      TEODORO SCHMIDT                            2433746
#> 192 09118               TOLTÉN                            1849038
#> 193 09119               VILCÚN                            3061660
#> 194 09120           VILLARRICA                            3542918
#> 195 09121             CHOLCHOL                            2060379
#> 196 09201                ANGOL                            5628638
#> 197 09202           COLLIPULLI                            2473847
#> 198 09203           CURACAUTÍN                            2424757
#> 199 09204              ERCILLA                            1531196
#> 200 09205            LONQUIMAY                            1915216
#> 201 09206           LOS SAUCES                            1503170
#> 202 09207               LUMACO                            1909557
#> 203 09208                PURÉN                            2110956
#> 204 09209              RENAICO                            1507673
#> 205 09210             TRAIGUÉN                            2380813
#> 206 09211             VICTORIA                            3369798
#> 207 10101         PUERTO MONTT                           10931180
#> 208 10102              CALBUCO                            2743014
#> 209 10103              COCHAMÓ                            1147708
#> 210 10104               FRESIA                            1468966
#> 211 10105            FRUTILLAR                            1478129
#> 212 10106          LOS MUERMOS                            1836056
#> 213 10107           LLANQUIHUE                            1575890
#> 214 10108              MAULLÍN                            1960491
#> 215 10109         PUERTO VARAS                            1002531
#> 216 10201               CASTRO                            4453910
#> 217 10202                ANCUD                            4350639
#> 218 10203              CHONCHI                            1897299
#> 219 10204      CURACO DE VÉLEZ                            1372554
#> 220 10205             DALCAHUE                            1764069
#> 221 10206            PUQUELDÓN                            1201034
#> 222 10207              QUEILÉN                            1472148
#> 223 10208              QUELLÓN                            2713327
#> 224 10209              QUEMCHI                            1952483
#> 225 10210             QUINCHAO                            1887735
#> 226 10301               OSORNO                            7508314
#> 227 10302         PUERTO OCTAY                            1000339
#> 228 10303            PURRANQUE                            1921759
#> 229 10304              PUYEHUE                            1015329
#> 230 10305            RÍO NEGRO                            1504726
#> 231 10306 SAN JUAN DE LA COSTA                            2147515
#> 232 10307            SAN PABLO                            1262905
#> 233 10401              CHAITÉN                            1523447
#> 234 10402            FUTALEUFÚ                            1256618
#> 235 10403            HUALAIHUÉ                            1417463
#> 236 10404               PALENA                             937883
#> 237 11101            COIHAIQUE                            4253003
#> 238 11102           LAGO VERDE                             942488
#> 239 11201                AISÉN                            2407488
#> 240 11202               CISNES                            1417018
#> 241 11203            GUAITECAS                            1023074
#> 242 11301             COCHRANE                            1118025
#> 243 11302            O´HIGGINS                             930990
#> 244 11303               TORTEL                            1064802
#> 245 11401          CHILE CHICO                            1791852
#> 246 11402           RÍO IBÁÑEZ                            1159512
#> 247 12101         PUNTA ARENAS                            4012832
#> 248 12102        LAGUNA BLANCA                             862316
#> 249 12103            RÍO VERDE                             756154
#> 250 12104         SAN GREGORIO                             786016
#> 251 12201       CABO DE HORNOS                            1919260
#> 252 12301             PORVENIR                            1292217
#> 253 12302            PRIMAVERA                             832130
#> 254 12303             TIMAUKEL                             799528
#> 255 12401              NATALES                            2093886
#> 256 12402     TORRES DEL PAINE                             886576
#> 257 13101             SANTIAGO                            1600755
#> 258 13102            CERRILLOS                            2550083
#> 259 13103          CERRO NAVIA                           12205331
#> 260 13104             CONCHALÍ                            7076391
#> 261 13105            EL BOSQUE                           14355667
#> 262 13106     ESTACIÓN CENTRAL                            2023086
#> 263 13107           HUECHURABA                            2214089
#> 264 13108        INDEPENDENCIA                            2034598
#> 265 13109          LA CISTERNA                            1771731
#> 266 13110           LA FLORIDA                           22613097
#> 267 13111            LA GRANJA                           10399881
#> 268 13112           LA PINTANA                           17835829
#> 269 13113             LA REINA                            1140916
#> 270 13114           LAS CONDES                            2545452
#> 271 13115         LO BARNECHEA                            1378424
#> 272 13116            LO ESPEJO                            8493987
#> 273 13117             LO PRADO                            8021835
#> 274 13118                MACUL                            2582458
#> 275 13119                MAIPÚ                           40617354
#> 276 13120                ÑUÑOA                            2412334
#> 277 13121  PEDRO AGUIRRE CERDA                            6027436
#> 278 13122            PEÑALOLÉN                           11181535
#> 279 13123          PROVIDENCIA                            1465687
#> 280 13124             PUDAHUEL                           12017301
#> 281 13125            QUILICURA                            5778407
#> 282 13126        QUINTA NORMAL                            2960731
#> 283 13127             RECOLETA                            3077062
#> 284 13128                RENCA                            5131736
#> 285 13129          SAN JOAQUÍN                            3180885
#> 286 13130           SAN MIGUEL                            1405043
#> 287 13131            SAN RAMÓN                            7615690
#> 288 13132             VITACURA                            1159430
#> 289 13201          PUENTE ALTO                           49849667
#> 290 13202               PIRQUE                             844487
#> 291 13203    SAN JOSÉ DE MAIPO                            1317334
#> 292 13301               COLINA                            3144210
#> 293 13302                LAMPA                            2369520
#> 294 13303               TILTIL                            1137925
#> 295 13401         SAN BERNARDO                           14660430
#> 296 13402                 BUIN                            2023952
#> 297 13403      CALERA DE TANGO                             821096
#> 298 13404                PAINE                            2869600
#> 299 13501            MELIPILLA                            7420794
#> 300 13502                ALHUÉ                             950380
#> 301 13503             CURACAVÍ                            2131286
#> 302 13504          MARÍA PINTO                            1177545
#> 303 13505            SAN PEDRO                            1165696
#> 304 13601            TALAGANTE                            4355439
#> 305 13602             EL MONTE                            2189897
#> 306 13603        ISLA DE MAIPO                            2357926
#> 307 13604        PADRE HURTADO                            3155463
#> 308 13605             PEÑAFLOR                            5609349
#> 309 14101             VALDIVIA                            9997817
#> 310 14102               CORRAL                            1560219
#> 311 14103                LANCO                            2086924
#> 312 14104            LOS LAGOS                            1787738
#> 313 14105                MÁFIL                            1192844
#> 314 14106            MARIQUINA                            1904213
#> 315 14107             PAILLACO                            2051117
#> 316 14108          PANGUIPULLI                            3711286
#> 317 14201             LA UNIÓN                            3344632
#> 318 14202              FUTRONO                            1882667
#> 319 14203           LAGO RANCO                            1610361
#> 320 14204            RÍO BUENO                            2671752
#> 321 15101                ARICA                           12957349
#> 322 15102            CAMARONES                            1132891
#> 323 15201                PUTRE                            1603124
#> 324 15202        GENERAL LAGOS                             955498
#> 325 16101              CHILLÁN                           11351767
#> 326 16102               BULNES                            2234391
#> 327 16103        CHILLÁN VIEJO                            2461738
#> 328 16104            EL CARMEN                            1937320
#> 329 16105               PEMUCO                            1388983
#> 330 16106                PINTO                            1711717
#> 331 16107              QUILLÓN                            3204681
#> 332 16108          SAN IGNACIO                            2112333
#> 333 16109               YUNGAY                            2038723
#> 334 16201             QUIRIHUE                            1942260
#> 335 16202           COBQUECURA                            1703395
#> 336 16203              COELEMU                            2386396
#> 337 16204               NINHUE                            1532461
#> 338 16205           PORTEZUELO                            1348892
#> 339 16206              RÁNQUIL                            1341122
#> 340 16207             TREGUACO                            1361438
#> 341 16301           SAN CARLOS                            4915689
#> 342 16302             COIHUECO                            2832761
#> 343 16303               ÑIQUÉN                            1950629
#> 344 16304           SAN FABIÁN                            1157915
#> 345 16305          SAN NICOLÁS                            1515440